+++ backend/canisters/user/impl/src/timer_job_types.rs
@@ -1,4 +1,3 @@
-use crate::model::token_swaps::TokenSwap;
 use crate::updates::end_video_call::end_video_call_impl;
 use crate::updates::swap_tokens::process_token_swap;
 use crate::{can_borrow_state, flush_pending_events, mutate_state, openchat_bot, read_state, run_regular_jobs};
@@ -9,6 +8,7 @@ use serde::{Deserialize, Serialize};
 use tracing::error;
 use types::{BlobReference, Chat, ChatId, CommunityId, EventIndex, MessageId, MessageIndex, P2PSwapStatus, UserId};
 use user_canister::C2CReplyContext;
+use user_state::TokenSwap;
 
 #[derive(Serialize, Deserialize, Clone)]
 pub enum TimerJob {